#576 make refactoring of features forms
