+++ apps/belis/desktop/src/components/ui/featuresForm/SchaltstelleForm.tsx
@@ -1,26 +1,14 @@
-import { useState, useEffect, useCallback } from "react";
-import {
-  Form,
-  Row,
-  Col,
-  Select,
-  Input,
-  DatePicker,
-  InputNumber,
-  message,
-} from "antd";
+import { useState, useEffect, useCallback, useRef } from "react";
+import type { FormInstance } from "antd";
+import { message } from "antd";
 import { useSelector } from "react-redux";
 import dayjs from "dayjs";
 import type { DraftFile } from "../../../store/slices/featuresForms";
-import { getKeyTablesData } from "../../../store/slices/keyTables";
 import { getJWT } from "../../../store/slices/auth";
 import { DokumentItem } from "../DocumentPreview";
 import { getDocumentKey } from "../FilePreview";
 import FeatureFormLayout from "./FeatureFormLayout";
-import StrassenschluesselFields from "./StrassenschluesselFields";
-import { getFormClassName, getPlaceholder } from "./readOnlyFormUtils";
-import { FormItem } from "./DraftFieldHighlight";
-import toTitleCase from "../../../helper/toTitleCase";
+import SchaltstelleFormFields from "./SchaltstelleFormFields";
 import { updateDataByClassName } from "../../../helper/apiMethods";
 import { uploadDraftFiles } from "../../../helper/uploadDraftFiles";
 
@@ -60,20 +48,6 @@ interface SchaltstelleFormProps {
   onRemovedDocumentKeysChange?: (keys: Set<string>) => void;
 }
 
-interface BauartItem {
-  id: number;
-  bezeichnung?: string;
-}
-
-interface RundsteuerempfaengerItem {
-  id: number;
-  rs_typ?: string;
-}
-
-const FormLabel = ({ children }: { children: React.ReactNode }) => (
-  <span className="text-sm font-medium text-gray-700">{children}</span>
-);
-
 const SchaltstelleForm = ({
   data,
   rawFeature,
@@ -93,22 +67,23 @@ const SchaltstelleForm = ({
   onRemovedDocumentKeysChange,
 }: SchaltstelleFormProps) => {
   const removedDocumentKeys = removedDocumentKeysProp ?? new Set<string>();
-  const [form] = Form.useForm();
+  const formRef = useRef<FormInstance | null>(null);
   const [saving, setSaving] = useState(false);
   const [localDocuments, setLocalDocuments] = useState<DokumentItem[] | null>(
     null
   );
-  const keyTablesData = useSelector(getKeyTablesData);
   const jwt = useSelector(getJWT);
 
-  // Key table options - sorted alphabetically
-  const bauartOptions = [
-    ...((keyTablesData.bauart || []) as BauartItem[]),
-  ].sort((a, b) => (a.bezeichnung || "").localeCompare(b.bezeichnung || ""));
-  const rundsteuerempfaengerOptions = [
-    ...((keyTablesData["rundsteuerempfänger"] ||
-      []) as RundsteuerempfaengerItem[]),
-  ].sort((a, b) => (a.rs_typ || "").localeCompare(b.rs_typ || ""));
+  const setFormInstance = useCallback((form: FormInstance) => {
+    formRef.current = form;
+  }, []);
+
+  const handleValuesChange = useCallback(
+    (_: Record<string, unknown>, allValues: Record<string, unknown>) => {
+      onDraftChange?.(allValues);
+    },
+    [onDraftChange]
+  );
 
   const handleToggleRemoveDocument = useCallback(
     (key: string) => {
@@ -161,62 +136,6 @@ const SchaltstelleForm = ({
     ? `ID - ${rawProps?.id || ""}`
     : "";
 
-  useEffect(() => {
-    // Reset form when data changes to clear old values
-    form.resetFields();
-
-    if (data) {
-      const schaltstelleData = data as Record<string, unknown>;
-      const { schaltstelle } = schaltstelleData;
-      if (
-        !schaltstelle ||
-        !Array.isArray(schaltstelle) ||
-        schaltstelle.length === 0
-      ) {
-        return;
-      }
-      const ss = schaltstelle[0];
-      const serverValues = {
-        // Strassenschluessel
-        strassenschluessel_pk: ss.tkey_strassenschluessel?.pk,
-        strassenschluessel_strasse: toTitleCase(
-          ss.tkey_strassenschluessel?.strasse || ""
-        ),
-        // Hausnummer
-        haus_nummer: ss.haus_nummer,
-        // Standortbez.
-        zusaetzliche_standortbezeichnung: ss.zusaetzliche_standortbezeichnung,
-        // Laufende Nr.
-        laufende_nummer: ss.laufende_nummer,
-        // Schaltstellen Nr.
-        schaltstellen_nummer: ss.schaltstellen_nummer,
-        // Bauart - use id for Select value
-        fk_bauart: ss.bauart?.id ?? null,
-        // Erstellungsjahr - parse as date string
-        erstellungsjahr: ss.erstellungsjahr
-          ? dayjs(ss.erstellungsjahr as string)
-          : null,
-        // Rundsteuerempfaenger - use id for Select value
-        rundsteuerempfaenger: ss.rundsteuerempfaengerObject?.id ?? null,
-        // Einbaudatum
-        einbaudatum_rs: ss.einbaudatum_rs
-          ? dayjs(ss.einbaudatum_rs as string)
-          : null,
-        // Pruefung
-        pruefdatum: ss.pruefdatum ? dayjs(ss.pruefdatum as string) : null,
-        // Bemerkung
-        bemerkung: ss.bemerkung,
-      };
-      form.setFieldsValue(serverValues);
-      onOriginalValues?.(form.getFieldsValue());
-
-      if (draftValues) {
-        form.setFieldsValue(draftValues);
-      }
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [data, form]);
-
   const handleSave = async () => {
     if (!jwt) {
       message.error("Nicht authentifiziert");
@@ -228,9 +147,13 @@ const SchaltstelleForm = ({
       return;
     }
 
+    if (!formRef.current) {
+      return;
+    }
+
     setSaving(true);
     try {
-      const formValues = form.getFieldsValue();
+      const formValues = formRef.current.getFieldsValue();
 
       // Remove display-only fields that the backend doesn't expect
       const { strassenschluessel_pk, strassenschluessel_strasse, ...rest } =
@@ -317,158 +240,14 @@ const SchaltstelleForm = ({
       onCancel={onCancel}
       onSave={handleSave}
     >
-      <Form
-        form={form}
-        layout="vertical"
-        requiredMark={false}
-        className={getFormClassName(readOnly, "pr-2")}
-        onValuesChange={(_, allValues) => onDraftChange?.(allValues)}
-      >
-        {/* Strassenschluessel - always disabled */}
-        <StrassenschluesselFields label="Strassenschlussel" />
-
-        {/* Hausnummer */}
-        <FormItem
-          name="haus_nummer"
-          label={<FormLabel>Hausnummer</FormLabel>}
-          className="mb-4"
-        >
-          <Input size="large" />
-        </FormItem>
-
-        {/* Standortbez. */}
-        <FormItem
-          name="zusaetzliche_standortbezeichnung"
-          label={<FormLabel>Standortbez.</FormLabel>}
-          className="mb-4"
-        >
-          <Input size="large" />
-        </FormItem>
-
-        {/* Laufende Nr. and Schaltstellen Nr. */}
-        <Row gutter={16}>
-          <Col span={12}>
-            <FormItem
-              name="laufende_nummer"
-              label={<FormLabel>Laufende Nr.</FormLabel>}
-              className="mb-4"
-            >
-              <InputNumber className="w-full" size="large" />
-            </FormItem>
-          </Col>
-          <Col span={12}>
-            <FormItem
-              name="schaltstellen_nummer"
-              label={<FormLabel>Schaltstellen Nr.</FormLabel>}
-              className="mb-4"
-            >
-              <Input size="large" />
-            </FormItem>
-          </Col>
-        </Row>
-
-        {/* Bauart */}
-        <FormItem
-          name="fk_bauart"
-          label={<FormLabel>Bauart</FormLabel>}
-          className="mb-4"
-        >
-          <Select
-            placeholder={getPlaceholder(readOnly, "Bauart auswählen")}
-            className="w-full"
-            size="large"
-            allowClear
-            showSearch
-            optionFilterProp="children"
-          >
-            {bauartOptions.map((item) => (
-              <Select.Option key={item.id} value={item.id}>
-                {item.bezeichnung}
-              </Select.Option>
-            ))}
-          </Select>
-        </FormItem>
-
-        {/* Erstellungsjahr */}
-        <FormItem
-          name="erstellungsjahr"
-          label={<FormLabel>Erstellungsjahr</FormLabel>}
-          className="mb-4"
-        >
-          <DatePicker
-            className="w-full"
-            size="large"
-            format="DD.MM.YYYY"
-            placeholder={getPlaceholder(readOnly, "Datum auswählen")}
-          />
-        </FormItem>
-
-        {/* Rundsteuerempfaenger */}
-        <FormItem
-          name="rundsteuerempfaenger"
-          label={<FormLabel>Rundsteuerempf.</FormLabel>}
-          className="mb-4"
-        >
-          <Select
-            placeholder={getPlaceholder(
-              readOnly,
-              "Rundsteuerempfänger auswählen"
-            )}
-            className="w-full"
-            size="large"
-            allowClear
-            showSearch
-            optionFilterProp="children"
-          >
-            {rundsteuerempfaengerOptions.map((item) => (
-              <Select.Option key={item.id} value={item.id}>
-                {item.rs_typ}
-              </Select.Option>
-            ))}
-          </Select>
-        </FormItem>
-
-        {/* Einbaudatum and Pruefung */}
-        <Row gutter={16}>
-          <Col span={12}>
-            <FormItem
-              name="einbaudatum_rs"
-              label={<FormLabel>Einbaudatum</FormLabel>}
-              className="mb-4"
-            >
-              <DatePicker
-                className="w-full"
-                size="large"
-                format="DD.MM.YYYY"
-                placeholder={getPlaceholder(readOnly, "Datum auswählen")}
-              />
-            </FormItem>
-          </Col>
-          <Col span={12}>
-            <FormItem
-              name="pruefdatum"
-              label={<FormLabel>Prüfung</FormLabel>}
-              className="mb-4"
-            >
-              <DatePicker
-                className="w-full"
-                size="large"
-                format="DD.MM.YYYY"
-                placeholder={getPlaceholder(readOnly, "Datum auswählen")}
-              />
-            </FormItem>
-          </Col>
-        </Row>
-
-        {/* Bemerkung */}
-        <FormItem
-          name="bemerkung"
-          label={<FormLabel>Bemerkung</FormLabel>}
-          className="mb-4"
-        >
-          <Input.TextArea rows={4} size="large" />
-        </FormItem>
-      </Form>
+      <SchaltstelleFormFields
+        schaltstelle={ss}
+        readOnly={readOnly}
+        onFormInstance={setFormInstance}
+        draftValues={draftValues}
+        onValuesChange={handleValuesChange}
+        onOriginalValues={onOriginalValues}
+      />
     </FeatureFormLayout>
   );
 };
